pyproject.toml
[project]
name = "lmcache-benchmark-orchestrator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.15.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tomli-w>=1.2.0",
]
